+++ .graphqlconfig
@@ -48,7 +48,8 @@
         "graphql-schemas/ApiDirective.graphql",
         "graphql-schemas/journey-planner-v3.graphql",
         "src/logic/use-stop-places-with-departures/GetStopPlacesWithDepartures.graphql",
-        "src/logic/use-nearest-places/GetNearestPlaces.graphql"
+        "src/logic/use-nearest-places/GetNearestPlaces.graphql",
+        "src/logic/use-walk-info/GetWalkInfo.graphql"
       ],
       "extensions": {
         "endpoints": {
