+++ src/dashboards/BusStop/DepartureTile/DepartureTile.tsx
@@ -13,7 +13,7 @@ import {
     IconColorType,
 } from '../../../types'
 import { useSettings } from '../../../settings/SettingsProvider'
-import { WalkInfo } from '../../../logic/useWalkInfo'
+import { WalkInfo } from '../../../logic/use-walk-info/useWalkInfo'
 import { Tile } from '../components/Tile/Tile'
 import { TileRows } from '../components/TileRows/TileRows'
 import './DepartureTile.scss'
