+++ src/dashboards/Chrono/BikeTile/BikeTile.tsx
@@ -6,7 +6,7 @@ import { Tile } from '../components/Tile/Tile'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { IconColorType } from '../../../types'
 import { getIconColorType, getTranslation } from '../../../utils'
-import { useWalkInfo, WalkInfo } from '../../../logic/useWalkInfo'
+import { useWalkInfo, WalkInfo } from '../../../logic/use-walk-info/useWalkInfo'
 import { TileRow } from '../components/TileRow/TileRow'
 import './BikeTile.scss'
