+++ src/logic/index.ts
@@ -2,6 +2,6 @@ export { useBikeRentalStations } from './useBikeRentalStations'
 export { useNearestPlaces } from './use-nearest-places/useNearestPlaces'
 export { useStopPlacesWithDepartures } from './use-stop-places-with-departures/useStopPlacesWithDepartures'
 export { useMobility } from './useMobility'
-export { useWalkInfo } from './useWalkInfo'
+export { useWalkInfo } from './use-walk-info/useWalkInfo'
 export { useWeather } from './useWeather'
 export { useOperators } from './useOperators'
