+++ src/logic/use-nearest-places/GetNearestPlaces.graphql
