+++ src/logic/use-stop-places-with-departures/GetStopPlacesWithDepartures.graphql
