+++ src/logic/use-walk-info/GetWalkInfo.graphql
@@ -1,0 +1,13 @@
+query getWalkInfo($from: Location!, $to: Location!) @api(name: journey_planner_v3) {
+    trip(
+        from: $from
+        to: $to
+        modes: { directMode: foot }
+        numTripPatterns: 1
+    ) {
+        tripPatterns {
+            duration
+            walkDistance
+        }
+    }
+}
